Extracted common components from LicensePage
