+++ app/components/LicensePage.js
@@ -6,77 +6,15 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import Consts from '../utils/consts';
-import StageUtils from '../utils/stageUtils';
 import Manager from '../utils/Manager';
-import {Button, Form, Grid, Header, Icon, Message, Segment, Table} from './basic';
+import {Button, Form, Grid, Header, Icon, Message, Segment} from './basic';
 import MessageContainer from './MessageContainer';
 import Banner from '../containers/banner/Banner';
 import FullScreenSegment from './layout/FullScreenSegment';
-import Logo from './banner/Logo';
+import EulaLink from './license/EulaLink';
+import CurrentLicense from './license/CurrentLicense';
+import UploadLicense from './license/UploadLicense';
 
-
-function CurrentLicense({license}) {
-    const formatTrial = (isTrial) => isTrial ? 'Yes' : 'No';
-    const formatExpirationDate = StageUtils.formatLocalTimestamp;
-    const formatVersion = (version) => _.isEmpty(version) ? 'All' : String(version);
-    const formatCapabilities = (capabilities) => _.join(capabilities, ', ');
-    const isFalse = (boolValue) => !boolValue;
-
-    const fields = [
-        {name: 'expiration_date', header: 'Expiration Date', icon: 'clock', format: formatExpirationDate},
-        {name: 'cloudify_version', header: 'Valid For Version', icon: 'thumbs up', format: formatVersion},
-        {name: 'license_edition', header: 'License Edition', icon: 'file alternate outline', format: String},
-        {name: 'capabilities', header: 'Capabilities', icon: 'wrench', format: formatCapabilities, hide: _.isEmpty},
-        {name: 'trial', header: 'Trial', icon: 'lab', format: formatTrial, hide: isFalse},
-        {name: 'customer_id', header: 'Licensed To', icon: 'handshake', format: String, hide: _.isEmpty},
-    ];
-
-    return !_.isEmpty(license)
-        ?
-        <Table basic='very' size='large' celled >
-            <Table.Body>
-                {
-                    _.map(fields, (field) => {
-                        const value = license[field.name];
-
-                        return !!field.hide && field.hide(value)
-                        ?
-                            null
-                        :
-                            <Table.Row key={field.header}>
-                                <Table.Cell width={5}>
-                                    <Header as='h4'>
-                                        <Icon name={field.icon} size='large'
-                                                   style={{display: 'inline-block', float: 'left'}}/>
-                                        <Header.Content>
-                                            {field.header}
-                                        </Header.Content>
-                                    </Header>
-                                </Table.Cell>
-                                <Table.Cell>
-                                    {field.format(license[field.name])}
-                                </Table.Cell>
-                            </Table.Row>
-                    })
-                }
-            </Table.Body>
-        </Table>
-        :
-        <Message>There is no license.</Message>
-}
-
-function UploadLicense({error, isLoading, license, onChange, onErrorDismiss, onUpload}) {
-    return (
-        <Form errors={error} errorMessageHeader='License error' onErrorsDismiss={onErrorDismiss}>
-            <Form.TextArea name='license' autoHeight placeholder='License in YAML format' error={!!error}
-                           value={license} onChange={onChange} disabled={isLoading} />
-
-            <Button content='Upload' icon='upload' color='yellow' labelPosition='left'
-                    disabled={_.isEmpty(license)} loading={isLoading} onClick={onUpload} />
-        </Form>
-
-    );
-}
 
 function LicenseSwitchButton({isEditLicenseActive, onClick, color}) {
     return (
@@ -221,7 +159,7 @@ export default class LicensePage extends Component {
 
                     <Grid columns={'equal'}>
                         <Grid.Column textAlign='left' verticalAlign='middle'>
-                            <a href='https://cloudify.co/license' target='_blank'> End User License Agreement</a>
+                            <EulaLink />
                         </Grid.Column>
 
                         <Grid.Column textAlign='right' verticalAlign='middle'>
