+++ app/components/license/EulaLink.js
@@ -1,0 +1,9 @@
+/**
+ * Created by jakub.niezgoda on 15/03/2019.
+ */
+
+import React from 'react';
+
+export default function EulaLink() {
+    return <a href='https://cloudify.co/license' target='_blank'>End User License Agreement</a>;
+}
